Initial commit of Amour Angels and Showy Beauty
Two sites by the same company, but different page coding.  Also the performers are first name only, and have different ids for the two sites, so I appended the performer ID to the name for each.  Also scenes don't have descriptions, so I pulled in the performer bio as the scene description